se realiza cambio en tipo de dato de int a long del campo IntNumero en TblDocumentos.
git-svn-id: http://192.168.0.251:18080/svn/HGInetMiFacturaElectronica/Tronco@659 30a10932-8c21-4e4d-906a-15261f94da71

+++ HGInetMiFacturaElectonicaData/ModeloServicio/Respuestas/DocumentoRespuesta.cs
@@ -50,7 +50,7 @@ namespace HGInetMiFacturaElectonicaData.ModeloServicio
         /// <summary>
         /// Número de Documento
         /// </summary>
-		public int Documento { get; set; }
+		public long Documento { get; set; }
 
         /// <summary>
         /// Código identificador del documento ante la DIAN
